Добавил update для departure, постарался по правилам не выходить за рамки назначения класса

+++ src/main/java/com/e_mail/item_post/db/repository/MailPostRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MailPostRepository extends JpaRepository<Post, Long> {
+public interface MailPostRepository extends JpaRepository<Post, Integer> {
     Optional<Post> findByName(String name);
 
 }
